s2: use an s2 configured version of UEISettings
IR does not work on s2 initially because the included version of
UEISettings is from an msm8996 LeEco device.

To fix this, we must use a version that is created by s2's LeRemote
app so it is already configured correctly without needing to ever run
the LeRemote app for CIR apps like MiRemote or Zaza Remote to work.

Change-Id: I531b960efefcb53f69600db6f5790d8030443d57
Signed-off-by: zhaoguomanong <zhaoguomanong@gmail.com>