s2: Move ACDB blobs to system
In our libacdbloader.so, the path "system" is specified, and acdb blobs were moved to "vendor",
apparently because of this the interlocutor hears an echo of his voice, so for now we return them back.

This solves the echo problem on the interlocutor's side and fixes volume control during a call.

This reverts commit 5f9381c705b0d2323d9267b0a36924c89a463474.

Change-Id: I88b3e19099d7a3a1d6da7855cbd72bfc3fc29c01